// src/api/notification/enum/notification-type.enum.ts
export enum NotificationType {
  COURSE_ENROLLED = 'COURSE_ENROLLED',
  COURSE_COMPLETED = 'COURSE_COMPLETED',
  NEW_LECTURE_ADDED = 'NEW_LECTURE_ADDED',
  COURSE_UPDATED = 'COURSE_UPDATED',
  COURSE_DISCOUNT_AVAILABLE = 'COURSE_DISCOUNT_AVAILABLE',
  //   INSTRUCTOR_ANNOUNCEMENT = 'INSTRUCTOR_ANNOUNCEMENT',

  NEW_ENROLLMENT = 'NEW_ENROLLMENT',
  COURSE_REVIEW_RECEIVED = 'COURSE_REVIEW_RECEIVED',
  COURSE_APPROVED = 'COURSE_APPROVED',
  COURSE_REJECTED = 'COURSE_REJECTED',
  PAYOUT_PROCESSED = 'PAYOUT_PROCESSED',
  NEW_COMMENT = 'NEW_COMMENT',

  //   USER_REPORTED_CONTENT = 'USER_REPORTED_CONTENT',
}